pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
"discord.py" = "^1.7.3"
Flask = "^2.0.2"
pafy = "^0.5.5"
youtube_dl = "^2021.12.17"
humanfriendly = "^10.0"
config = "^0.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"